Move selectSiteIfLoggedIn to controller-logged-in.jsx to fix docker builds
